Rescope and move models to their own package.
