+++ core/src/main/java/com/novoda/noplayer/Player.java
@@ -2,6 +2,11 @@ package com.novoda.noplayer;
 
 import android.net.Uri;
 
+import com.novoda.noplayer.model.Bitrate;
+import com.novoda.noplayer.model.PlayerAudioTrack;
+import com.novoda.noplayer.model.PlayerSubtitleTrack;
+import com.novoda.noplayer.model.Timeout;
+import com.novoda.noplayer.model.VideoPosition;
 import com.novoda.noplayer.player.PlayerInformation;
 
 import java.util.List;
