+++ core/src/main/java/com/novoda/noplayer/exoplayer/mediasource/ExoPlayerAudioTrackSelector.java
@@ -5,7 +5,7 @@ import com.google.android.exoplayer2.source.TrackGroup;
 import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.trackselection.MappingTrackSelector;
 import com.google.android.exoplayer2.trackselection.TrackSelection;
-import com.novoda.noplayer.PlayerAudioTrack;
+import com.novoda.noplayer.model.PlayerAudioTrack;
 import com.novoda.noplayer.exoplayer.RendererTypeRequester;
 
 import java.util.ArrayList;
