+++ core/src/main/java/com/novoda/noplayer/model/Timeout.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer;
+package com.novoda.noplayer.model;
 
 import java.util.concurrent.TimeUnit;
