+++ core/src/test/java/com/novoda/noplayer/exoplayer/ExoPlayerTwoImplTest.java
@@ -7,16 +7,16 @@ import com.google.android.exoplayer2.ExoPlayerLibraryInfo;
 import com.google.android.exoplayer2.text.Cue;
 import com.novoda.noplayer.ContentType;
 import com.novoda.noplayer.Heart;
-import com.novoda.noplayer.LoadTimeout;
+import com.novoda.noplayer.model.LoadTimeout;
 import com.novoda.noplayer.Player;
 import com.novoda.noplayer.Player.StateChangedListener;
 import com.novoda.noplayer.PlayerListenersHolder;
-import com.novoda.noplayer.PlayerSubtitleTrack;
+import com.novoda.noplayer.model.PlayerSubtitleTrack;
 import com.novoda.noplayer.PlayerView;
 import com.novoda.noplayer.SurfaceHolderRequester;
-import com.novoda.noplayer.TextCues;
-import com.novoda.noplayer.Timeout;
-import com.novoda.noplayer.VideoPosition;
+import com.novoda.noplayer.model.TextCues;
+import com.novoda.noplayer.model.Timeout;
+import com.novoda.noplayer.model.VideoPosition;
 import com.novoda.noplayer.exoplayer.drm.DrmSessionCreator;
 import com.novoda.noplayer.exoplayer.forwarder.ExoPlayerForwarder;
 import com.novoda.noplayer.exoplayer.mediasource.ExoPlayerTrackSelector;
