+++ core/src/test/java/com/novoda/noplayer/mediaplayer/LoadTimeoutTest.java
@@ -3,9 +3,9 @@ package com.novoda.noplayer.mediaplayer;
 import android.os.Handler;
 
 import com.novoda.noplayer.Clock;
-import com.novoda.noplayer.LoadTimeout;
+import com.novoda.noplayer.model.LoadTimeout;
 import com.novoda.noplayer.Player;
-import com.novoda.noplayer.Timeout;
+import com.novoda.noplayer.model.Timeout;
 
 import org.junit.Before;
 import org.junit.Test;
